Add ascii art to cli interface and print the proj version
